+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 0.3.2 (Unreleased)
 * Remove all puts and replace with logger. This is required because OpenStudio PAT reads the result from the command line.
 * Add method to list status of all instances in the group
 * Add method to `delete_key_pair`
+* Add launch time to the server data struct
 
 Version 0.3.1
 -------------